Cut constants.js and pasted to Enums.js
