+++ src/Enums.js
@@ -65,6 +65,12 @@ const userFilterType = {
   ALL: 2
 };
 
+const DEFAULT_PICS = {
+  EVENT: 'https://i.gyazo.com/640f22609f95f72a28afa0a130e557a1.png',
+  INVENTORY: 'https://user-images.githubusercontent.com/25803515/9'+
+    '9110346-1ce92000-259f-11eb-97df-7ed5c2284ef3.png'
+};
+
 module.exports = {
   eventModalState,
   memberApplicationState,
@@ -73,5 +79,6 @@ module.exports = {
   membershipState,
   membershipStateToString,
   itemCategories,
-  userFilterType
+  userFilterType,
+  DEFAULT_PICS
 };
