--- src/util/constants.js
@@ -1,9 +1,0 @@
-const DEFAULT_PICS = {
-  EVENT: 'https://i.gyazo.com/640f22609f95f72a28afa0a130e557a1.png',
-  INVENTORY: 'https://user-images.githubusercontent.com/25803515/9'+
-    '9110346-1ce92000-259f-11eb-97df-7ed5c2284ef3.png'
-};
-
-module.exports = {
-  DEFAULT_PICS
-};
